refactor: função get
