+++ server/bin/www
@@ -12,7 +12,7 @@ var http = require('http');
  * Get port from environment and store in Express.
  */
 
-var port = normalizePort(process.env.PORT || '3300');
+var port = normalizePort(process.env.PORT || '3330');
 app.set('port', port);
 
 /**n dev
